refactor: use order notification
